Fix an interesting race condition in GTServer, revealed by unit tests for the UofC NSD.  On open, the NSD created a NewGTConnection which creates a new GTServer and then:
    server.Start();
    server.StartSeparateListeningThread(...);

Turns out that the server listening thread never had a chance to run
until after test case was finished and the server instance had been
shutdown.  Since the server wasn't active, the listening thread proceeded
to start the server, which includes starting the acceptors.  But
the next unit test had been started -- which also opened a new shared
dictiony using the same host and port, resulting in a bind class.

We now track the listening thread and explicitly abort it in Stop().