Two major changes to LightweightDotNetSerializingMarshaller and DotNetSerializingMarshaller to their internal programming conventions and their on-wire formats.
To fix <https://papyrus.usask.ca/trac/gt/ticket/59>, which complains
of an asymmetry between the Marshal*() methods, which are responsible
for encoding the marshalled-payload length, and the Unmarshal*()
methods, which have no details of the length.  We now expose the length
to the Unmarshal* methods (which were previously using Stream.Position
and Stream.Length to calculate length information), and have changed
the Marshal*() conventions to remove the responsibility for calculating
and encoding the marshalled-payload length, and move this upwards.

To support more efficient marshalling, we now encode the payload
length using a fixed-size encoding, versus the adaptive encoding
used previously.  This helps avoid unnecessary byte-array copies.
It unfortunately breaks interoperability with previous versions.